requirements.txt
# for document_processor
requests==2.32.3
beautifulsoup4==4.12.3
pymupdf==1.25.1
python-pptx==1.0.2
langchain==0.3.14
openai==1.59.7
anthropic==0.43.1
langchain-anthropic
langchain-openai
langchain-community
tqdm==4.67.1
qdrant-client==1.12.2
python-dotenv==1.0.1
numpy==2.2.1
torch==2.5.1
# torchvision
transformers==4.48.0
aiohttp==3.11.11
httpx==0.28.1

openpyxl

# for the app.py

streamlit==1.41.1
Pillow==11.1.0 # Image library for handling images
opencv-python==4.11.0.86
